+++ Source/Steam Library Manager/Definitions/App.cs
@@ -91,18 +91,21 @@ namespace Steam_Library_Manager.Definitions
 
                         System.Diagnostics.Process.Start(string.Format(action, AppId, Properties.Settings.Default.SteamID64));
                         break;
+
                     case "google":
                         {
-                            var url = string.Format(@"https://www.google.com/search?q={0}+pc", System.Uri.EscapeUriString(AppName));
+                            var url = string.Format(@"https://www.google.com/search?q={0}+pc", Uri.EscapeUriString(AppName));
                             System.Diagnostics.Process.Start(url);
                         }
                         break;
+
                     case "youtube":
                         {
-                            var url = string.Format(@"https://www.youtube.com/results?search_query={0}+gameplay+pc", System.Uri.EscapeUriString(AppName));
+                            var url = string.Format(@"https://www.youtube.com/results?search_query={0}+pc", Uri.EscapeUriString(AppName));
                             System.Diagnostics.Process.Start(url);
                         }
                         break;
+
                     case "compress":
                         if (Functions.TaskManager.TaskList.Count(x => x.App == this && x.TargetLibrary == Library && x.TaskType == Enums.TaskType.Compress && !x.Completed) == 0)
                         {